+++ framework/tests/bin/setup_funds.rs
