+++ framework/tests/four-chain-starship/configs/four-junos.yaml
