+++ framework/tests/four-chain-starship/configs/two-junos.yaml
