+++ framework/tests/src/clone_testing/common.rs
