+++ framework/tests/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod clone_testing;
+pub mod interchain;
