--- interchain/framework-clone-testing/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "abstract-framework-clone-testing"
-version = { workspace = true }
-edition = { workspace = true }
-
-[lib]
-
-
-[dependencies]
-cosmwasm-std = { workspace = true }
-cosmwasm-schema = { workspace = true }
-
-cw-orch = { workspace = true, features = ["daemon"] }
-cw-orch-clone-testing = { workspace = true }
-
-abstract-std = { workspace = true }
-abstract-app = { workspace = true, features = ["test-utils"] }
-abstract-unit-test-utils = { workspace = true }
-abstract-client = { workspace = true }
-abstract-interface = { workspace = true, features = ["daemon"] }
-abstract-integration-test-utils = { path = "../../framework/packages/abstract-integration-test-utils" }
-
-tokio = { workspace = true }
-anyhow = { workspace = true }
-env_logger = "0.11.3"
-lazy_static = "1.4.0"
-
-cw2 = { workspace = true }
-semver = { workspace = true }
