+++ modules/contracts/adapters/dex/Cargo.toml
@@ -25,7 +25,7 @@ required-features = ["schema"]
 default = ["export"]
 export = []
 schema = ["abstract-adapter/schema"]
-testing = ["dep:abstract-client", "dep:cw20", "dep:serde_json"]
+testing = ["dep:abstract-client", "dep:cw20", "dep:serde_json", "export"]
 
 # Supported Dexes
 # wynd = ["abstract-wyndex-adapter/full_integration"]
